+++ src/commands/branch.js
@@ -19,20 +19,14 @@ class BranchCommand extends AbstractAuthenticatedCommand {
     return null;
   }
 
-  async createBranch(branchName) {
+  async createBranch(branchName, environmentSecret) {
     try {
-      await BranchManager.createBranch(branchName);
-      // FIXME: Handle branch creation
-      //        Cases: #1, #2, #4
+      await BranchManager.createBranch(branchName, environmentSecret);
+
       return this.log(`✅ Switched to new branch: ${branchName}.`);
-    } catch (err) {
-      const customError = BranchManager.handleError(err);
-      // FIXME: Display correct error, depending on the case
-      //        - Branch already exist
-      //        - Server branch creation failed
-      //        - Absence of a remote or a production environment
-      //        Cases: #5, #6, #7
-      // return this.error(`❌ Failed create ${branchName}.`);
+    } catch (error) {
+      const customError = BranchManager.handleError(error);
+
       return this.error(customError);
     }
   }
@@ -76,7 +70,8 @@ class BranchCommand extends AbstractAuthenticatedCommand {
       if (config.delete) {
         return this.deleteBranch(config.BRANCH_NAME, config.force);
       }
-      return this.createBranch(config.BRANCH_NAME);
+      // TODO: Replace process.env.FOREST_ENV_SECRET if --project
+      return this.createBranch(config.BRANCH_NAME, process.env.FOREST_ENV_SECRET);
     }
     return this.listBranches();
   }