+++ src/services/branch-manager.js
@@ -1,12 +1,12 @@
-const P = require('bluebird');
-const agent = require('superagent-promise')(require('superagent'), P);
+const agent = require('superagent');
 const authenticator = require('./authenticator');
 const { serverHost } = require('../config');
+const ApiErrorDeserializer = require('../deserializers/api-error');
 
 const ERROR_MESSAGE_PROJECT_IN_V1 = '⚠️  This project does not support branches yet. Please migrate your environments from your Project settings first.';
 const ERROR_MESSAGE_ENV_SECRET_ISSUE = '⚠️  Your development environment is not properly set up. Please run `forest init` first and retry.';
 const ERROR_MESSAGE_BRANCH_ALREADY_EXISTS = '❌ This branch already exists.';
-const ERROR_MESSAGE_NO_PRODUCTION_OR_REMOTE_ENVIRONMENT = '❌ You cannot create a branch until this project has either a remote or a production environment.';
+const ERROR_MESSAGE_NO_PRODUCTION_OR_REMOTE_ENVIRONMENT = '❌ You cannot run branch commands until this project has either a remote or a production environment.';
 
 async function getBranches() {
   // FIXME: Implement getBranches function
@@ -16,32 +16,30 @@ async function deleteBranch() {
   // FIXME: Implement deleteBranch function
 }
 
-async function createBranch(branchName) {
+async function createBranch(branchName, environmentSecret) {
   const authToken = authenticator.getAuthToken();
 
   await agent
     .post(`${serverHost()}/api/branches`)
     .set('Authorization', `Bearer ${authToken}`)
-    .set('forest-secret-key', `${process.env.FOREST_ENV_SECRET}`)
+    .set('forest-secret-key', `${environmentSecret}`)
     .send({ branchName });
 }
 
 function handleError(error) {
   try {
-    const { status } = error;
+    const apiError = ApiErrorDeserializer.deserialize(error);
 
-    if (status === 404) {
+    if (apiError.status === 404) {
       // NOTICE: When no env/project can be found through envSecret
       return ERROR_MESSAGE_ENV_SECRET_ISSUE;
     }
 
-    const message = JSON.parse(error.response.text).errors[0].detail;
-
-    if (!message) {
+    if (!apiError.message) {
       return 'Unknown server error';
     }
 
-    switch (message) {
+    switch (apiError.message) {
       case 'Workflow disabled.':
         return ERROR_MESSAGE_PROJECT_IN_V1;
       case 'Not development environment.':
@@ -51,9 +49,9 @@ function handleError(error) {
       case 'No production/remote environment.':
         return ERROR_MESSAGE_NO_PRODUCTION_OR_REMOTE_ENVIRONMENT;
       default:
-        return message;
+        return apiError.message;
     }
-  } catch (err) {
+  } catch (_) {
     // NOTICE: Client issue or not well formatted server answer
     return 'Oops something went wrong';
   }